Import live.2005.06.04.tar.gz
- Updated "MPEG2TransportStreamFramer::doStopGettingFrames()" to clear the PID status
  table.  This makes it possible to handle PCR discontinuities when seeking (e.g.)
  within a stream.
- Updated the "ByteStreamFileSource::seekToByte*()" operations to take 64-bit
  parameters, to allow for seeking within files that are >4 GB in size.

+++ liveMedia/include/ByteStreamFileSource.hh
@@ -45,8 +45,8 @@ public:
   unsigned fileSize() const { return fFileSize; }
       // 0 means zero-length, unbounded, or unknown
 
-  void seekToByteAbsolute(unsigned byteNumber);
-  void seekToByteRelative(int offset);
+  void seekToByteAbsolute(u_int64_t byteNumber);
+  void seekToByteRelative(int64_t offset);
 
 protected:
   ByteStreamFileSource(UsageEnvironment& env,
